test: fix  passed parameter to assertEqual. Now if more explicitly

+++ src/test/java/internal/infrastructure/aggregation/java/aggregators/AvgAggregatorTest.java
@@ -82,7 +82,7 @@ public class AvgAggregatorTest {
         Number result = aggregator.finish();
 
         // assert
-        assertEquals(2.0, result);
+        assertEquals(2.0, result.intValue());
     }
 
     @Test
